move shared logic into a liv
